Revert "Solve exogenous impact block from shock Jacobian"
This reverts commit 0726f4f21948fd5bf449fbd4c0852b6fef488444.
