+++ SymbolicDSGE/_symbolic_printers/__init__.py
@@ -1,26 +1,21 @@
 """Symbolic expression printers used by native numeric callbacks."""
 
-from .base import ExpressionPrinter, Layout
-from .ops import (
-    OpTable,
-    ConstraintOpTable,
-    ConstraintOps,
-    F64Ops,
-    C128Ops,
-    BicomplexOps,
-)
-
+from .base import ExpressionPrinter, Layout, OpTable
 from .constraint_printer import (
     ConstraintLayout,
+    ConstraintOps,
     ConstraintPrinter,
     build_constraint_cfunc,
 )
 from .measurement_printer import (
+    F64Ops,
     MeasurementLayout,
     MeasurementPrinter,
     build_measurement_cfunc,
 )
 from .residual_printer import (
+    BicomplexOps,
+    C128Ops,
     ResidualLayout,
     ResidualPrinter,
     build_cfunc,
@@ -31,7 +26,6 @@ __all__ = [
     "BicomplexOps",
     "C128Ops",
     "ConstraintLayout",
-    "ConstraintOpTable",
     "ConstraintOps",
     "ConstraintPrinter",
     "ExpressionPrinter",
