+++ SymbolicDSGE/core/solver_backend.py
@@ -28,9 +28,7 @@ class KleinSolution:
 
     ``steady_state`` is the Newton-resolved steady state the solve linearized at
     (the seed after convergence), so second-order and measurement callers reuse
-    it instead of re-solving. ``a``/``b`` are the pencil at that point, kept for
-    the same reason: the state-space assembly and the second-order sweep both
-    need them, and recomputing means another complex-step jacobian.
+    it instead of re-solving.
     """
 
     p: NDC
@@ -38,8 +36,6 @@ class KleinSolution:
     stab: int
     eig: NDC
     steady_state: NDF
-    a: NDF
-    b: NDF
     order: int = 1
 
 
@@ -56,8 +52,7 @@ class PerturbationSolution:
     * ``hxx`` (nx, nx, nx), ``gxx`` (ny, nx, nx) -- the state-quadratic terms;
     * ``hss`` (nx,), ``gss`` (ny,) -- the sigma^2 risk correction.
 
-    ``steady_state`` is the (nonlinear) expansion point the tensors are taken at,
-    and ``a``/``b`` the pencil there.
+    ``steady_state`` is the (nonlinear) expansion point the tensors are taken at.
     """
 
     p: NDC
@@ -70,8 +65,6 @@ class PerturbationSolution:
     hxx: NDF
     gss: NDF
     hss: NDF
-    a: NDF
-    b: NDF
 
 
 def klein_solve(
@@ -106,11 +99,5 @@ def klein_solve(
         n_states,
     )
     return KleinSolution(
-        p=p,
-        f=f,
-        stab=stab,
-        eig=eig,
-        steady_state=np.asarray(ss, dtype=float64),
-        a=a,
-        b=b,
+        p=p, f=f, stab=stab, eig=eig, steady_state=np.asarray(ss, dtype=float64)
     )
