--- tests/_oracles/dynare_post82_first_order.py
@@ -1,1049 +1,0 @@
-"""Dynare first-order goldens for the POST82 New Keynesian model.
-
-Generated by ``tests/fixtures/models/make_post82_first_order_goldens.m`` after
-``dynare post82_first_order.mod noclearall``, then transcribed here at ``%.17g``
-and the generated artifacts deleted. Regenerate by rerunning that pair; only the
-``.mod`` and the ``.m`` are kept in the tree.
-
-Everything here is raw. Unlike ``dynare_rbc_first_order``, no array below has
-been re-dated, rescaled, or reordered on the way out of Dynare. There is no
-``SDSGE`` view and no seeding transform: the ``.mod`` is written at Dynare's own
-dating and the yaml is written the same way, so a comparison against these is
-``ours == dynare`` with nothing in between.
-
-Ordering is Dynare's declaration order ``[g, z, r, x, Pi]`` throughout, shocks
-``[e_g, e_z, e_r]``. ``ORDER_VAR`` and ``STATE_VAR`` are the identity here, so
-the raw DR-ordered ``ghx``/``ghu`` and the declaration-ordered ones coincide;
-both are kept anyway so a reordering bug cannot hide behind a fixture that
-happens not to permute.
-
-Paths follow ``simult_``: one period per row of the shock matrix, initial
-condition dropped, so row 0 is ``x1`` and carries the first innovation. IRF row 0
-is the impact response, matching ``oo_.irfs``. Every path is the same length as
-the shock block that drove it.
-
-This model exists to pin ``B``. ``e_r`` enters a Taylor rule carrying ``Pi(t)``
-and ``x(t)``, so the impact of ``e_r`` on ``r`` is a fixed point, ``0.3336``, not
-one, and ``e_g`` moves ``r`` on impact by ``0.3470`` through the same rule.
-``sdsge_assemble_state_space`` writes those two entries as ``1.0`` and ``0.0``.
-See https://github.com/GongJr0/SymbolicDSGE/issues/390.
-
-Filter conventions, recorded both ways rather than chosen:
-
-* ``P0``: ``P0_UNCOND`` is the unconditional covariance (doubling iteration);
-  ``P0_FIXED`` is the identity, which is the yaml's own P0.
-* the ``n * log(2 * pi)`` term: ``*_CONST`` includes it, ``*_NOCONST`` omits it.
-  The gap is ``12 * 3 * 0.5 * log(2 * pi) == 33.0818``, the signature to look for
-  if our loglik is off by a constant.
-
-The recursion is predict-then-update, the observation at ``t`` paired to the
-state at ``t``, three observables and unit measurement standard deviations.
-"""
-
-from __future__ import annotations
-
-import numpy as np
-
-DECL_COLUMNS = ("g", "z", "r", "x", "Pi")
-EXO_COLUMNS = ("e_g", "e_z", "e_r")
-
-# 1-based Dynare indices, exactly as dumped. Identity for this model.
-ORDER_VAR = (1, 2, 3, 4, 5)
-STATE_VAR = (1, 2, 3)
-
-YS = np.array([0.0, 0.0, 0.0, 0.0, 0.0])
-
-# Rows in DR order, columns following STATE_VAR.
-GHX_RAW = np.array(
-    [
-        [0.82999999999999996, 1.51791636030829e-17, 0.0],
-        [0.0, 0.84999999999999976, 0.0],
-        [0.28802956547406666, -0.04720983377331555, 0.28018450737268485],
-        [0.8915474645568261, 0.70755859893091522, -1.7109927961514928],
-        [0.6998723949067891, -0.11521874007419936, -1.3632643790298022],
-    ]
-)
-
-# Rows in DR order, one column per shock in declaration order.
-GHU_RAW = np.array(
-    [
-        [0.99999999999999989, -1.7383576652802728e-17, 0.0],
-        [-0.0, 1.0, -0.0],
-        [0.34702357286032132, -0.055540980909783016, 0.33355298496748226],
-        [1.0741535717552129, 0.83242188109519466, -2.0368961858946375],
-        [0.84321975289974604, -0.13555145891082293, -1.6229337845592895],
-    ]
-)
-
-# Same arrays with rows permuted into declaration order. Identity here.
-GHX_DECL = GHX_RAW.copy()
-GHU_DECL = GHU_RAW.copy()
-
-# x_t = A x_{t-1} + R e_t, declaration order, deviations from YS.
-A_DECL = np.array(
-    [
-        [0.82999999999999996, 1.51791636030829e-17, 0.0, 0.0, 0.0],
-        [0.0, 0.84999999999999976, 0.0, 0.0, 0.0],
-        [0.28802956547406666, -0.04720983377331555, 0.28018450737268485, 0.0, 0.0],
-        [0.8915474645568261, 0.70755859893091522, -1.7109927961514928, 0.0, 0.0],
-        [0.6998723949067891, -0.11521874007419936, -1.3632643790298022, 0.0, 0.0],
-    ]
-)
-
-R_DECL = GHU_RAW.copy()
-
-SIGMA_E = np.array(
-    [
-        [0.032399999999999998, 0.041471999999999995, 0.0],
-        [0.041471999999999995, 0.40960000000000002, 0.0],
-        [0.0, 0.0, 0.032399999999999998],
-    ]
-)
-
-# Initial deviation on [g, z, r]; x and Pi are not free and start at YS.
-X0_STATES = np.array([0.5, -0.3, 0.2])
-
-# Innovations driving STOCH, one row per period, columns EXO_COLUMNS.
-SHOCK_BLOCK = np.array(
-    [
-        [0.25, -0.10, 0.40],
-        [-0.70, 0.55, -0.15],
-        [0.10, 0.20, 0.00],
-        [0.60, -0.35, 0.25],
-        [-0.20, 0.05, -0.50],
-        [0.00, 0.30, 0.10],
-        [0.35, -0.45, 0.20],
-        [-0.45, 0.15, -0.30],
-        [0.15, 0.60, 0.05],
-        [-0.10, -0.25, 0.35],
-        [0.05, 0.40, -0.20],
-        [0.20, -0.05, 0.15],
-    ]
-)
-
-# Zero-innovation decay from X0_STATES.
-DET = np.array(
-    [
-        [
-            0.41499999999999998,
-            -0.25499999999999989,
-            0.21421463434356497,
-            -0.10869240663116009,
-            0.11184894366969389,
-        ],
-        [
-            0.34444999999999998,
-            -0.21674999999999986,
-            0.1915903990795047,
-            -0.17695494112836629,
-            0.027796642137762062,
-        ],
-        [
-            0.28589349999999997,
-            -0.18423749999999983,
-            0.16312517688133552,
-            -0.17407959478849916,
-            0.0048563419075332015,
-        ],
-        [
-            0.23729160499999996,
-            -0.15660187499999981,
-            0.13674874965125036,
-            -0.1545772048271605,
-            -0.001066161807556143,
-        ],
-        [
-            0.19695203214999996,
-            -0.13311159374999978,
-            0.11405502742099957,
-            -0.13322440000259766,
-            -0.0023073846630307093,
-        ],
-        [
-            0.16347018668449995,
-            -0.11314485468749978,
-            0.094968636124968309,
-            -0.1137394981547553,
-            -0.0023089155885930301,
-        ],
-        [
-            0.13568025494813496,
-            -0.096173126484374788,
-            0.079034537149333894,
-            -0.096805836659885569,
-            -0.0020226801004991912,
-        ],
-        [
-            0.11261461160695201,
-            -0.081747157511718543,
-            0.065764495047523561,
-            -0.082310259050334275,
-            -0.0017051577741532958,
-        ],
-        [
-            0.093470127633770161,
-            -0.069485083884960747,
-            0.054721800012254213,
-            -0.069962210054199975,
-            -0.0014197310831679938,
-        ],
-        [
-            0.077580205936029226,
-            -0.059062321302216621,
-            0.045534740086074277,
-            -0.059460318909958856,
-            -0.0011771348146335653,
-        ],
-        [
-            0.064391570926904257,
-            -0.050202973106884111,
-            0.037891844095446874,
-            -0.050533229670044352,
-            -0.0009745583956158807,
-        ],
-        [
-            0.053445003869330529,
-            -0.042672527140851485,
-            0.031533457880896364,
-            -0.042946075795165885,
-            -0.00080639504529650408,
-        ],
-    ]
-)
-
-# Same start, driven by SHOCK_BLOCK.
-STOCH = np.array(
-    [
-        [
-            0.66499999999999992,
-            -0.35499999999999987,
-            0.4399458196366165,
-            -0.7381546761597314,
-            -0.31296448603800314,
-        ],
-        [
-            -0.1480499999999999,
-            0.24825000000000025,
-            0.0080681665276108872,
-            -0.39958940452749647,
-            -0.51481173102119526,
-        ],
-        [
-            -0.022881499999999916,
-            0.41101250000000017,
-            -0.028507882034646173,
-            0.30375297864462925,
-            -0.086006520812732551,
-        ],
-        [
-            0.58100835499999992,
-            -0.00063937499999994207,
-            0.27705988608622178,
-            0.16311270432380237,
-            0.12313472370942341,
-        ],
-        [
-            0.28223693464999988,
-            0.049456531250000053,
-            0.0060474004094267231,
-            0.88873513396234816,
-            0.66504487205139229,
-        ],
-        [
-            0.23425665575949989,
-            0.3420380515625,
-            0.097345139157068647,
-            0.32231090465385298,
-            -0.019371501527496476,
-        ],
-        [
-            0.54443302428038476,
-            -0.1592676561718751,
-            0.29176217197326237,
-            -0.1217092733529837,
-            0.023371725759183816,
-        ],
-        [
-            0.001879410152719363,
-            0.014622492253906189,
-            -0.018478603002992672,
-            0.12605673896925906,
-            0.088733814774914918,
-        ],
-        [
-            0.15155991042675704,
-            0.61242911841582026,
-            0.030080178613200389,
-            0.60236996523530195,
-            -0.011172818240927279,
-        ],
-        [
-            0.025794725654208331,
-            0.27056475065344709,
-            0.11909549097161257,
-            -0.51144911858635089,
-            -0.62395888510264086,
-        ],
-        [
-            0.071409622292992914,
-            0.62998003805543001,
-            -0.04355077252776772,
-            0.8047217792466056,
-            0.13704740718864553,
-        ],
-        [
-            0.25926998650318411,
-            0.48548303234711537,
-            0.10083928921889741,
-            0.45160311125515329,
-            -0.031255210175124233,
-        ],
-    ]
-)
-
-# Unit-innovation responses from the steady state, row 0 is impact.
-IRF_EG = np.array(
-    [
-        [
-            0.99999999999999989,
-            0.0,
-            0.34702357286032132,
-            1.0741535717552129,
-            0.84321975289974604,
-        ],
-        [
-            0.82999999999999985,
-            0.0,
-            0.38526019428264474,
-            0.29779263129806355,
-            0.22678751934265973,
-        ],
-        [
-            0.68889999999999985,
-            0.0,
-            0.34700847708886295,
-            0.080806978520635897,
-            0.055682588249004183,
-        ],
-        [
-            0.57178699999999982,
-            0.0,
-            0.29564996686237316,
-            0.020458043830652528,
-            0.0090777968146607546,
-        ],
-        [
-            0.47458320999999981,
-            0.0,
-            0.24752810147380472,
-            0.0039202866326056851,
-            -0.0028711314182466863,
-        ],
-        [
-            0.3939040642999998,
-            0.0,
-            0.20604753492992159,
-            -0.00040534086999599594,
-            -0.0052985557828607588,
-        ],
-        [
-            0.32694037336899984,
-            0.0,
-            0.17118734354849391,
-            -0.0013616781245751053,
-            -0.0052146839117023225,
-        ],
-        [
-            0.27136050989626986,
-            0.0,
-            0.14213253519797547,
-            -0.0014174506653899521,
-            -0.0045570654989172399,
-        ],
-        [
-            0.22522922321390398,
-            0.0,
-            0.11798318410831932,
-            -0.0012569692436176227,
-            -0.0038464923923696992,
-        ],
-        [
-            0.1869402552675403,
-            0.0,
-            0.097929735612012556,
-            -0.0010658351758900952,
-            -0.0032105563656760205,
-        ],
-        [
-            0.15516041187205845,
-            0.0,
-            0.081282715223909743,
-            -0.00089096155379236874,
-            -0.002669796049083073,
-        ],
-        [
-            0.12878314185380851,
-            0.0,
-            0.067464943533211599,
-            -0.00074126839561627911,
-            -0.0022173412439582263,
-        ],
-        [
-            0.10689000773866106,
-            0.0,
-            0.055995984357316446,
-            -0.0006157487806672246,
-            -0.0018407884392386692,
-        ],
-        [
-            0.08871870642308867,
-            0.0,
-            0.046476689774489446,
-            -0.00051121046291732752,
-            -0.0015279651353775903,
-        ],
-        [
-            0.07363652633116359,
-            0.0,
-            0.038575658889241943,
-            -0.00042434362285308225,
-            -0.0012682420874207054,
-        ],
-        [
-            0.061118316854865774,
-            0.0,
-            0.032017798664643701,
-            -0.00035221611696939492,
-            -0.0010526496252995926,
-        ],
-    ]
-)
-
-IRF_EZ = np.array(
-    [
-        [
-            -1.7383576652802728e-17,
-            1.0,
-            -0.055540980909783016,
-            0.83242188109519466,
-            -0.13555145891082293,
-        ],
-        [
-            7.5079498125663564e-19,
-            0.84999999999999976,
-            -0.062771556148518801,
-            0.80258881715874153,
-            -0.039501699223517928,
-        ],
-        [
-            1.3525448897063469e-17,
-            0.72249999999999959,
-            -0.057715976243807775,
-            0.70882648946461235,
-            -0.012361702549524581,
-        ],
-        [
-            2.2193068287790069e-17,
-            0.61412499999999948,
-            -0.050280227272625322,
-            0.6099627073035917,
-            -0.0045634051894955507,
-        ],
-        [
-            2.7742150526609037e-17,
-            0.52200624999999945,
-            -0.043080479875004539,
-            0.52055853122076967,
-            -0.0022134659377747307,
-        ],
-        [
-            3.0949603207667286e-17,
-            0.44370531249999939,
-            -0.036714311322288778,
-            0.44306040160406279,
-            -0.0014148187907534818,
-        ],
-        [
-            3.2423246192358358e-17,
-            0.37714951562499938,
-            -0.031234035278824859,
-            0.37676543143880276,
-            -0.0010718542441919554,
-        ],
-        [
-            3.2636108540152772e-17,
-            0.32057708828124937,
-            -0.026556458728201289,
-            0.32029659221990886,
-            -0.00087444430092485648,
-        ],
-        [
-            3.1954062158747834e-17,
-            0.2724905250390619,
-            -0.022575099355615393,
-            0.27226498500887997,
-            -0.00073301399108900711,
-        ],
-        [
-            3.0658049851618581e-17,
-            0.23161694628320256,
-            -0.019189425483740096,
-            0.23142884648845027,
-            -0.00062018617258865691,
-        ],
-        [
-            2.8961932897722619e-17,
-            0.19687440434072212,
-            -0.016311177259039505,
-            0.19671553076576509,
-            -0.00052635251455448312,
-        ],
-        [
-            2.7026793097857089e-17,
-            0.16734324368961376,
-            -0.013864547068138521,
-            0.16720848448764652,
-            -0.00044717388370694305,
-        ],
-        [
-            2.4972368745056601e-17,
-            0.14224175713617165,
-            -0.011784878007905888,
-            0.14212729120106618,
-            -0.0003800345484716687,
-        ],
-        [
-            2.2886176961156912e-17,
-            0.12090549356574587,
-            -0.010017149949115277,
-            0.12080821976379229,
-            -0.00032301164378014183,
-        ],
-        [
-            2.083077114510618e-17,
-            0.10276966953088396,
-            -0.008514578477290707,
-            0.10268699303133548,
-            -0.00027455493166538697,
-        ],
-        [
-            1.8849497677682182e-17,
-            0.087354219101251337,
-            -0.0072373919916373592,
-            0.087283945822776376,
-            -0.00023337030064603595,
-        ],
-    ]
-)
-
-IRF_ER = np.array(
-    [
-        [0.0, 0.0, 0.33355298496748226, -2.0368961858946375, -1.6229337845592895],
-        [0.0, 0.0, 0.093456378775802568, -0.57070675441418928, -0.45472090292523165],
-        [0.0, 0.0, 0.026185029448133283, -0.15990319083980345, -0.12740575217816849],
-        [0.0, 0.0, 0.0073366395764644697, -0.044802396752770743, -0.035697117910486506],
-        [0.0, 0.0, 0.0020556127455026406, -0.012552937463290647, -0.010001779396374307],
-        [
-            0.0,
-            0.0,
-            0.00057595084444766955,
-            -0.0035171385992322098,
-            -0.0028023436330234042,
-        ],
-        [
-            0.0,
-            0.0,
-            0.00016137250362245213,
-            -0.0009854477457873316,
-            -0.00078517327030764249,
-        ],
-        [
-            0.0,
-            0.0,
-            4.5214075430953555e-05,
-            -0.00027610719119494626,
-            -0.0002199933859433467,
-        ],
-        [
-            0.0,
-            0.0,
-            1.2668283450933135e-05,
-            -7.7360957347011729e-05,
-            -6.1638738465785532e-05,
-        ],
-        [
-            0.0,
-            0.0,
-            3.5494567579572364e-06,
-            -2.1675341724151768e-05,
-            -1.7270219572109879e-05,
-        ],
-        [
-            0.0,
-            0.0,
-            9.9450279316889544e-07,
-            -6.0730949431160642e-06,
-            -4.8388479630297069e-06,
-        ],
-        [
-            0.0,
-            0.0,
-            2.7864427518478607e-07,
-            -1.701587114864518e-06,
-            -1.3557702327727982e-06,
-        ],
-        [
-            0.0,
-            0.0,
-            7.8071808974868117e-08,
-            -4.7675834753002311e-07,
-            -3.7986581477999673e-07,
-        ],
-        [
-            0.0,
-            0.0,
-            2.1874511337317781e-08,
-            -1.3358030273851481e-07,
-            -1.0643251618185693e-07,
-        ],
-        [
-            0.0,
-            0.0,
-            6.128899183064592e-09,
-            -3.7427131317484876e-08,
-            -2.9820742114848896e-08,
-        ],
-        [
-            0.0,
-            0.0,
-            1.7172225983438033e-09,
-            -1.0486502350562287e-08,
-            -8.3553099389368131e-09,
-        ],
-    ]
-)
-
-IRF = {"e_g": IRF_EG, "e_z": IRF_EZ, "e_r": IRF_ER}
-
-# --- Kalman -----------------------------------------------------------------
-# OutGap = x, Infl = 4*Pi + pi_star, Rate = 4*r + (r_star + pi_star).
-KF_Z = np.array(
-    [
-        [0.0, 0.0, 0.0, 1.0, 0.0],
-        [0.0, 0.0, 0.0, 0.0, 4.0],
-        [0.0, 0.0, 4.0, 0.0, 0.0],
-    ]
-)
-
-KF_D = np.array([0.0, 3.4300000000000002, 6.4399999999999995])
-
-KF_H = np.eye(3)
-
-# Observables implied by STOCH, one row per period.
-KF_DATA = np.array(
-    [
-        [-0.7381546761597314, 2.1781420558479878, 8.1997832785464659],
-        [-0.39958940452749647, 1.3707530759152191, 6.4722726661104435],
-        [0.30375297864462925, 3.0859739167490701, 6.3259684718614144],
-        [0.16311270432380237, 3.922538894837694, 7.5482395443448862],
-        [0.88873513396234816, 6.0901794882055693, 6.4641896016377061],
-        [0.32231090465385298, 3.3525139938900144, 6.8293805566282737],
-        [-0.1217092733529837, 3.5234869030367353, 7.607048687893049],
-        [0.12605673896925906, 3.7849352590996599, 6.3660855879880289],
-        [0.60236996523530195, 3.3853087270362909, 6.5603207144528008],
-        [-0.51144911858635089, 0.93416445958943672, 6.9163819638864501],
-        [0.8047217792466056, 3.9781896287545822, 6.2657969098889286],
-        [0.45160311125515329, 3.3049791592995033, 6.8433571568755891],
-    ]
-)
-
-P0_UNCOND = np.array(
-    [
-        [
-            0.10414657666345219,
-            0.14082173174872653,
-            0.036901499683624786,
-            0.17668780204601217,
-            0.026975464411906094,
-        ],
-        [
-            0.14082173174872653,
-            1.4760360360360338,
-            -0.043463051413674229,
-            1.4431590825074836,
-            -0.030971286914784284,
-        ],
-        [
-            0.036901499683624786,
-            -0.043463051413674229,
-            0.024283676401585449,
-            -0.051892559100681537,
-            -0.0070942283163670785,
-        ],
-        [
-            0.17668780204601217,
-            1.4431590825074836,
-            -0.051892559100681537,
-            1.5929501287986096,
-            0.11418645474500669,
-        ],
-        [
-            0.026975464411906094,
-            -0.030971286914784284,
-            -0.0070942283163670785,
-            0.11418645474500669,
-            0.11537919869690036,
-        ],
-    ]
-)
-
-P0_FIXED = np.eye(5)
-
-LOGLIK_UNCOND_CONST = -50.317487376073821
-LOGLIK_UNCOND_NOCONST = -17.235700180705599
-LOGLIK_FIXED_CONST = -51.324766476891497
-LOGLIK_FIXED_NOCONST = -18.242979281523279
-
-FILTERED_UNCOND = np.array(
-    [
-        [
-            0.11830038018180529,
-            -0.34137212832679642,
-            0.13524740710692904,
-            -0.61652314123753404,
-            -0.22003266330391183,
-        ],
-        [
-            0.011734425491455563,
-            -0.086134458186184315,
-            0.083641241606187397,
-            -0.51578331621091578,
-            -0.34230380032185381,
-        ],
-        [
-            0.016473306407067317,
-            0.13417141259242349,
-            0.013215910489702189,
-            0.038079527118544509,
-            -0.077043470054963412,
-        ],
-        [
-            0.13220569721095826,
-            0.049798236815215752,
-            0.052104268683505303,
-            0.12855356776948176,
-            0.060978657789815795,
-        ],
-        [
-            0.21393641579059847,
-            0.10330410340500513,
-            0.012557024680646574,
-            0.65743871302580748,
-            0.43860464688570977,
-        ],
-        [
-            0.17856332738638658,
-            0.17464787706650856,
-            0.071288633205054741,
-            0.22101768132829971,
-            0.034335823574314514,
-        ],
-        [
-            0.21317747221865574,
-            0.0010876671228163826,
-            0.11006859849064048,
-            0.0091262654034055257,
-            0.003709941661438055,
-        ],
-        [
-            0.15604219536230635,
-            0.025806397696267358,
-            0.066493421926301088,
-            0.10496715018615373,
-            0.061049311445484075,
-        ],
-        [
-            0.14565202964362794,
-            0.26082916463306427,
-            0.051468461059999561,
-            0.27946591390568437,
-            0.01247973314072681,
-        ],
-        [
-            0.034608735533252138,
-            0.18202742266427852,
-            0.086175340057100369,
-            -0.32594212466851102,
-            -0.40554068032895052,
-        ],
-        [
-            0.076932573506486898,
-            0.37399652685933593,
-            -0.0038326049602249584,
-            0.453548436798488,
-            0.06165322097489577,
-        ],
-        [
-            0.09779189009197925,
-            0.37085564416975619,
-            0.022832901170133646,
-            0.35577215691082215,
-            -0.014006345453887532,
-        ],
-    ]
-)
-
-FILTERED_FIXED = np.array(
-    [
-        [
-            0.45832209562617193,
-            -0.22348671734571765,
-            0.32336159337723741,
-            -0.6213314032154027,
-            -0.32232541436963136,
-        ],
-        [
-            0.16141603055740766,
-            -0.018242635150624908,
-            0.15830270278411041,
-            -0.46024826967207744,
-            -0.35417784487571896,
-        ],
-        [
-            0.10217188675404197,
-            0.17162352338376805,
-            0.054933241268985758,
-            0.075461245214599765,
-            -0.078258106112455206,
-        ],
-        [
-            0.20253008333885858,
-            0.075065258892343917,
-            0.086564861475601079,
-            0.15514486420426682,
-            0.06109418943382354,
-        ],
-        [
-            0.2603771342301705,
-            0.11905976316572914,
-            0.035369184182682908,
-            0.67420363024785224,
-            0.43879018279601234,
-        ],
-        [
-            0.21004942262213019,
-            0.18477697517828728,
-            0.086793829691264146,
-            0.23187370956167791,
-            0.034496734299701709,
-        ],
-        [
-            0.23544953140604519,
-            0.00790330608665335,
-            0.12106160437575425,
-            0.016479083934511723,
-            0.0038427803724969843,
-        ],
-        [
-            0.17083089892835671,
-            0.030218409621184272,
-            0.073801037211680728,
-            0.10974314435760039,
-            0.061143558689341154,
-        ],
-        [
-            0.15562692443687506,
-            0.26373826441221543,
-            0.056402249710144653,
-            0.28262477271726005,
-            0.012546828296808638,
-        ],
-        [
-            0.041368266987300575,
-            0.18396412310284269,
-            0.089521258943871973,
-            -0.32383396783362223,
-            -0.40549338487401054,
-        ],
-        [
-            0.081494807909704814,
-            0.37528655007393646,
-            -0.0015730897786520176,
-            0.4549552710427488,
-            0.061686044994918213,
-        ],
-        [
-            0.10088699150213784,
-            0.37172181525780301,
-            0.024366450024413254,
-            0.35671814755131476,
-            -0.013983602331740622,
-        ],
-    ]
-)
-
-PREDICTED_UNCOND = np.array(
-    [
-        [0.0, 0.0, 0.0, 0.0, 0.0],
-        [
-            0.098189315550898382,
-            -0.29016630907777685,
-            0.08808435666602113,
-            -0.36747772008389934,
-            -0.062250335546533364,
-        ],
-        [
-            0.009739573157908115,
-            -0.073214289458256643,
-            0.030881235003985444,
-            -0.19359294110830505,
-            -0.095888221078794467,
-        ],
-        [
-            0.013672844317865874,
-            0.11404570070355993,
-            0.0021134825707682155,
-            0.087008543628102106,
-            -0.021946628712784348,
-        ],
-        [
-            0.10973072868509536,
-            0.042328501292933376,
-            0.050326991891719734,
-            0.063952796452381874,
-            0.015757534331968581,
-        ],
-        [
-            0.17756722510619671,
-            0.087808487894254342,
-            0.060261327103607244,
-            0.24234319697386003,
-            0.12070707858459773,
-        ],
-        [
-            0.14820756173070085,
-            0.14845069550653223,
-            0.063160390927362392,
-            0.16079695113626619,
-            0.0076635809500263663,
-        ],
-        [
-            0.17693730194148424,
-            0.00092451705439392494,
-            0.092189582153232419,
-            0.0025008439827220066,
-            -0.00098089118568855937,
-        ],
-        [
-            0.12951502215071425,
-            0.021935438041827248,
-            0.062356876646240414,
-            0.043608796329207208,
-            0.015588130794387256,
-        ],
-        [
-            0.12089118460421119,
-            0.22170478993810458,
-            0.044059054710436502,
-            0.22634544992176128,
-            0.0017203074795130863,
-        ],
-        [
-            0.028725250492599278,
-            0.1547233092646367,
-            0.025519849892656127,
-            0.0122050125199272,
-            -0.11423104312660405,
-        ],
-        [
-            0.063854036010384127,
-            0.31789704783043543,
-            0.0034287053203137831,
-            0.3397710588785996,
-            0.01597642967063605,
-        ],
-    ]
-)
-
-PREDICTED_FIXED = np.array(
-    [
-        [0.0, 0.0, 0.0, 0.0, 0.0],
-        [
-            0.38040733936972271,
-            -0.18996370974385995,
-            0.23316199357045292,
-            -0.30278340311944213,
-            -0.094310501097027322,
-        ],
-        [
-            0.13397530536264834,
-            -0.015506239878031168,
-            0.091707785710377562,
-            -0.13985246466062384,
-            -0.10073591848947096,
-        ],
-        [
-            0.084802666005854829,
-            0.14587999487620279,
-            0.036717649278229451,
-            0.11853440625307285,
-            -0.023155494103334215,
-        ],
-        [
-            0.16809996917125261,
-            0.063805470058492311,
-            0.079044966573385486,
-            0.085566397322890009,
-            0.015085497788782903,
-        ],
-        [
-            0.21611302141104149,
-            0.10120079869086974,
-            0.079285418649953115,
-            0.25586391372374501,
-            0.12029530369555214,
-        ],
-        [
-            0.17434102077636807,
-            0.15706042890154415,
-            0.076095440057663305,
-            0.16950595049007436,
-            0.0073950858468950498,
-        ],
-        [
-            0.19542311106701749,
-            0.0067178101736553456,
-            0.10136289843911285,
-            0.0083709519602894328,
-            -0.0011649544585590477,
-        ],
-        [
-            0.14178964611053607,
-            0.025685648178006625,
-            0.068455650747516666,
-            0.047412107363331485,
-            0.015467778107332478,
-        ],
-        [
-            0.12917034728260629,
-            0.22417752475038305,
-            0.048177152348803133,
-            0.22885522380965601,
-            0.0016402198556979553,
-        ],
-        [
-            0.034335661599459479,
-            0.15636950463741625,
-            0.028312838129399727,
-            0.013876941586462133,
-            -0.11428474987730268,
-        ],
-        [
-            0.067640690565054992,
-            0.31899356756284591,
-            0.0053149430795885848,
-            0.34088526021334731,
-            0.015940460178174135,
-        ],
-    ]
-)
