--- tests/fixtures/models/rbc_lagged.yaml
@@ -1,48 +1,0 @@
-# Canonical RBC, identical economics to rbc_second_order.yaml, authored with a
-# lag on capital instead of relabelling it. This is a verbatim transcription of
-# rbc_second_order.mod, so it is the Dynare-parity spelling. Offsets of -1, 0
-# and 1 all compile directly; no equation is re-dated on the way in.
-#
-# rbc_second_order.yaml writes k(t+1) on the left of the resource constraint,
-# so its k(t) is the stock available for production at t. This file writes
-# k(t-1) on the right, so its k(t) is the stock chosen at t, which is Dynare's
-# dating. The two differ only in what the k column of a path means.
-
-name: "RBC_LAGGED"
-
-variables:
-  c: {ss_seed: c_ss}
-  k: {ss_seed: k_ss}
-  z: {ss_seed: null} # z_ss = 0
-
-shock_map:
-  e: z
-
-observables: [c_obs]
-
-equations:
-  model:
-    euler: "c(t)**(-gamma) = beta * c(t+1)**(-gamma) * (alpha * exp(z(t+1)) * k(t)**(alpha - 1) + 1 - delta)"
-    capital: "k(t) = exp(z(t)) * k(t-1)**alpha + (1 - delta) * k(t-1) - c(t)"
-    z_process: "z(t) = rho * z(t-1) + e"
-  constraint: {}
-  observables:
-    c_obs: c(t)
-
-calibration:
-  parameters:
-    beta: 0.99
-    gamma: 2.0
-    alpha: 0.33
-    delta: 0.025
-    rho: 0.95
-    sig: 0.01
-    # Analytic steady state, exact to machine precision:
-    #   k_ss = (alpha / (1/beta - 1 + delta))^(1/(1-alpha))
-    #   c_ss = k_ss^alpha - delta*k_ss
-    k_ss: 28.348419061048435
-    c_ss: 2.3066172319875173
-  shocks:
-    std:
-      e: sig
-    corr: {}
